before adding page CRUD

+++ public/delete_subject.php
@@ -7,6 +7,13 @@
 		redirect_to("manage_content.php");
 	}
 	$id = $current_subject["id"];
+	
+	$pages_set = find_pages_for_subject($id);
+	if ($pages_set && $pages_set->num_rows) {
+		$_SESSION["message"] = "Subject has pages: cannot delete";
+		redirect_to("manage_content.php?subject={$id}");
+	}
+	
 	$query = "DELETE FROM subjects where id={$id} LIMIT 1";
 	$result = $dblink->query($query);
 	if ($result && $dblink->affected_rows == 1) {
@@ -18,3 +25,4 @@
 		redirect_to("manage_content.php?subject={$id}");
 	}
 ?>
+<?php include("../includes/layout/footer.php"); ?>
